Initial check-in of some testing files
The primary purpose of this initial check-in was a proof-of-concept to
see how difficult it is to interact with Stockfish given a PGN file.
It's not that bad, even at this rudimentary, introductory level.

What this script currently does is actually interesting, though needing
some fine tuning. It will go through a PGN and find interesting moves to
look at based on evaluation swings. Current threshold set to 125
centipawns.