Add arguments and restructure a little
This will be renamed now to some other filename and become the beginning
of an analysis script that will provide places to look in a given PGN
file.

This is similar, obviously, to what Game Review or Computer Analysis or
Tactical Analysis are doing in chess.com, lichess.org or Fritz/CB. The
difference is intead of all the gobbedly-gook that is printed out from
those, this will be more bare bones - guiding to certain parts of the
game for the player to then look at without all the hints and variations
provided by the other analyses.

And it will help me learn about chess analysis with an engine and chess
programming. I will be adding more functionality as time goes on.