Use more functionality from python-chess
This module provides a lot of functions that I am not taking advantage
of, so this is a first step in using it more. I may move to using it
exclusively as it can also use an engine to perform evaluations.

This commit is mainly for converting the printed move to SAN instead of
the UCI. That is, Nf3 instead of g1f3.