Commit with all testing and debugging information
Spent a while working on getting the right evaluation for the right
position. There's something still weird about it, where you have to call
board.push(move) before doing the evaluation; otherwise, it gives the
wrong FEN or evaluation for the wrong position. It's a one-off error,
and I can't really figure it out. I have it giving the "correct" output
now, but I'm pretty sure it's a hack.

Will continue to look at it. I may just return to using the Stockfish
module; it was much easier to deal with. I want to use mostly pure
python-chess because it should be more universal for other engines.