Remove a lot of old code used for testing
I'm not sure I'm happy with how this is written. I may try the Stockfish
class again, or re-write this from scratch. As mentioned in the previous
commit, there is some hackery happening - we push the move before we
actually evaluate the position. Somehow it works as I expect it to. It
points me to the correct moves, but it really shouldn't. It should be
looking at the next position. I don't understand and it's annoying.