Intermediate commit before refactoring
I began adding categories to evaluations, like Inaccuracy, Mistake and
Blunder; however, I want to refactor to use classes so that different
engines and classes can be used without much hassle.

This commit has the intial code for assigning categories (though not
quite right as it assigns a ton of mistakes and blunders to Giri and
Nepo's game in the recent Superbet 2023 tournament).

I'm actually wondering if the Stockfish class isn't better at evaluating
than the python-chess. /shrug We'll see.