Refactor into initial class structure
Trying to make it more general and easier to manage. Maybe. Want to be
able to use different classes without a hassle. Need to improve the
configuration, including adding a config file.

This commit should be functionally equivalent to the last commit, though
there are probably a few bugs with the re-implementation I haven't found.