Add debug logging and hash size option
Also add some comments about comparing evaluation scores from move to
move (versus comparing to best move, I think).

I was trying to use the engine.analysis function instead of
engine.analyse as it seems to be faster when testing in the python3
interpreter directly, but it's not working correctly in the script.