First attempt at two pass analysis
Don't believe it's working as intended. About to re-work it again based
on the python-chess-annotator. There doesn't seem to be a way around
calling engine.analyse twice for each move: once for the engine move and
once for the player's move. That's overhead.

One of the things the chess-annotator does is classify the opening and
then set the root to the last move of the opening, so that no analysis
is done of the opening because there's really no need. Not a bad idea.