New algorithm inspired from python-chess-annotator
Using that module as a guide to move more toward doing annotations. It's
very old and uses an outdated version of python chess. So I am going to
just use it as a guide and udpate/change it how I want it.

The main thing this change does is analyse the game in reverse order,
then track the player's move/eval and the engine's move/eval.
Unfortunately, I haven't figured out how to do that without running
analyse twice for every ply. Need to think about it more...