Primarily fix bug in engine vs player evaluation
Since scores are normalized for White, Black was never considered
as being worse, so need to invert the sign for Black so a positive score
was good for Black.

Also added options for evaluating player moves, from move to move; and
computer vs player moves. Computer vs player is the default.